Move `EditorFeedbackPanel` and `MonacoBatchSpecEditor`
