+++ client/web/src/enterprise/batches/create/CreateOrEditBatchChangePage.tsx
@@ -55,14 +55,14 @@ import {
     CreateBatchSpecFromRawVariables,
     CreateBatchSpecFromRawResult,
 } from '../../../graphql-operations'
+import { EditorFeedbackPanel } from '../batch-spec/edit/editor/EditorFeedbackPanel'
+import { MonacoBatchSpecEditor } from '../batch-spec/edit/editor/MonacoBatchSpecEditor'
+import { ExecutionOptions, RunBatchSpecButton } from '../batch-spec/edit/RunBatchSpecButton'
 import { BatchSpecDownloadLink } from '../BatchSpec'
 
 import { GET_BATCH_CHANGE_TO_EDIT, CREATE_EMPTY_BATCH_CHANGE, CREATE_BATCH_SPEC_FROM_RAW } from './backend'
 import { CodeInsightsBatchesIcon } from './CodeInsightsBatchesIcon'
 import { DownloadSpecModal } from './DownloadSpecModal'
-import { EditorFeedbackPanel } from './editor/EditorFeedbackPanel'
-import { MonacoBatchSpecEditor } from './editor/MonacoBatchSpecEditor'
-import { ExecutionOptions, RunBatchSpecButton } from '../batch-spec/edit/RunBatchSpecButton'
 import { getTemplateRenderer } from './go-checker-templates'
 import { LibraryPane } from './library/LibraryPane'
 import { NamespaceSelector } from './NamespaceSelector'
